Remove exposure of ports in Dockerfile
As they cant utilize the env variables anyway (will only get default values).
And setting them to the same as default is misleading if they are changed, exposed
ports are not changed (or removed once exposed).

Further more it was disapointing and understandable that -P did not map to same
ports on host (security reason really).